fix bps

+++ common.sh
@@ -69,8 +69,9 @@ save_endpoint_to_file() {
   filename=$2
 
   response=$(curl -s -w "%{http_code}" "$endpoint" -o "$filename")
+  
 
-  if [ "${response}" -ne 200 ]; then
+  if [[ "${response}" -ne 200 && "${response}" -ne 301 ]]; then
     echo "Error: Failed to get file ${filename}, status code: $response"
     rm "$filename" 2>/dev/null
     return 1
